requirements.txt
streamlit
groq
google-generativeai
requests
python-dotenv